+++ kok-api/src/main/java/com/kok/kokapi/vote/application/service/CandidateService.java
@@ -3,9 +3,9 @@ package com.kok.kokapi.vote.application.service;
 import com.kok.kokcore.station.domain.entity.Station;
 import com.kok.kokcore.station.usecase.DeleteRecommendStationUseCase;
 import com.kok.kokcore.station.usecase.GetRecommendStationUseCase;
-import com.kok.kokcore.vote.application.port.out.LoadCandidatePort;
-import com.kok.kokcore.vote.application.port.out.SaveCandidatePort;
 import com.kok.kokcore.vote.domain.Candidate;
+import com.kok.kokcore.vote.port.out.LoadCandidatePort;
+import com.kok.kokcore.vote.port.out.SaveCandidatePort;
 import com.kok.kokcore.vote.usecase.GetCandidateUseCase;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
